improve zshrc and vs code config
- add vs code keybindings
- add toggle term vs code keybinding

+++ .zshrc
@@ -73,12 +73,17 @@ export FZF_DEFAULT_OPTS='--height 40% --layout=reverse --border'
 # Enable open vs code from terminal
 code () { VSCODE_CWD="$PWD" open -n -b "com.microsoft.VSCode" --args $* ;}
 
-ln -s \
-  $HOME/.config/vscode/settings.json \
-  $HOME/Library/Application\ Support/Code/User/settings.json
-ln -s \
-  $HOME/.config/vscode/keybindings.json \
-  $HOME/Library/Application\ Support/Code/User/keybindings.json
+if [ ! -L "$HOME/Library/Application Support/Code/User/settings.json" ]; then
+ ln -s \
+   $HOME/.config/vscode/settings.json \
+   $HOME/Library/Application\ Support/Code/User/settings.json
+fi
+if [ ! -L "$HOME/Library/Application Support/Code/User/keybindings.json" ]; then
+  ln -s \
+ ln -s \
+   $HOME/.config/vscode/keybindings.json \
+   $HOME/Library/Application\ Support/Code/User/keybindings.json
+fi
 
 # Enable zoxide shell
 eval "$(zoxide init zsh)"
